Create ha_day1p1.cpp
Q. a) Define a class to represent a bank account. Include the following members.
Data members
a) Name of the depositor.
b) Account number
c) Balance amount in the account.
Member functions.
a) To assign initial values
b) To deposit an amount.
c) To withdraw an amount after checking the balance
d) To display name and balance.
Write a program in C++ to test the program for any number of customers.