Create ha_day1p2.cpp
Q. Raising a number n to the power p is the same as multiplying n by itself p times.
Write a function called power () that takes a double value for n and an integer value for
p,and that returns the result as a double value. Use a default argument of 2 for p, so
that if this argument is omitted, the number will be squared. Write a main () function
that gets values from the user to test the function.